Fix kafka dependencies by copying the kafka project and common to each consumer
