+++ libraries/src/AWS.Lambda.Powertools.Kafka.Json/AWS.Lambda.Powertools.Kafka.Json.csproj
@@ -1,6 +1,10 @@
 ﻿<!--suppress MsbuildTargetFrameworkTagInspection -->
 <Project Sdk="Microsoft.NET.Sdk">
-
+    <Import Project="../KafkaDependencies.props" />
+    <PropertyGroup>
+        <IncludeKafkaFiles>true</IncludeKafkaFiles>
+        <DefineConstants>$(DefineConstants);KAFKA_JSON</DefineConstants>
+    </PropertyGroup>
     <PropertyGroup>
         <PackageId>AWS.Lambda.Powertools.Kafka.Json</PackageId>
         <Description>Powertools for AWS Lambda (.NET) - Kafka Json consumer package.</Description>
@@ -10,10 +14,10 @@
         <TargetMultiFramework>false</TargetMultiFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-    </PropertyGroup>    
+    </PropertyGroup>
 
-    <ItemGroup>
-      <ProjectReference Include="..\AWS.Lambda.Powertools.Kafka\AWS.Lambda.Powertools.Kafka.csproj" />
+    <!-- Debug: Use project references -->
+    <ItemGroup Condition="'$(IncludeKafkaFiles)' != 'true'">
+        <ProjectReference Include="..\AWS.Lambda.Powertools.Kafka\AWS.Lambda.Powertools.Kafka.csproj"/>
     </ItemGroup>
-
 </Project>
