+++ libraries/src/AWS.Lambda.Powertools.Kafka/ConsumerRecord.cs
@@ -13,7 +13,15 @@
  * permissions and limitations under the License.
  */
 
+#if KAFKA_JSON
+namespace AWS.Lambda.Powertools.Kafka.Json;
+#elif KAFKA_AVRO
+namespace AWS.Lambda.Powertools.Kafka.Avro;
+#elif KAFKA_PROTOBUF
+namespace AWS.Lambda.Powertools.Kafka.Protobuf;
+#else
 namespace AWS.Lambda.Powertools.Kafka;
+#endif
 
 /// <summary>
 /// Represents a single record consumed from a Kafka topic.
