+++ libraries/src/AWS.Lambda.Powertools.Kafka/ConsumerRecords.cs
@@ -15,7 +15,15 @@
 
 using System.Collections;
 
+#if KAFKA_JSON
+namespace AWS.Lambda.Powertools.Kafka.Json;
+#elif KAFKA_AVRO
+namespace AWS.Lambda.Powertools.Kafka.Avro;
+#elif KAFKA_PROTOBUF
+namespace AWS.Lambda.Powertools.Kafka.Protobuf;
+#else
 namespace AWS.Lambda.Powertools.Kafka;
+#endif
 
 /// <summary>
 /// Represents a collection of Kafka consumer records that can be enumerated.
