+++ libraries/src/AWS.Lambda.Powertools.Kafka/PowertoolsKafkaSerializerBase.cs
@@ -23,7 +23,15 @@ using System.Text.Json.Serialization;
 using System.Text.Json.Serialization.Metadata;
 using AWS.Lambda.Powertools.Common;
 
+#if KAFKA_JSON
+namespace AWS.Lambda.Powertools.Kafka.Json;
+#elif KAFKA_AVRO
+namespace AWS.Lambda.Powertools.Kafka.Avro;
+#elif KAFKA_PROTOBUF
+namespace AWS.Lambda.Powertools.Kafka.Protobuf;
+#else
 namespace AWS.Lambda.Powertools.Kafka;
+#endif
 
 /// <summary>
 /// Base class for Kafka event serializers that provides common functionality
