+++ libraries/tests/AWS.Lambda.Powertools.Kafka.Tests/PowertoolsKafkaSerializerBaseTests.cs
@@ -17,6 +17,7 @@ using System.Runtime.Serialization;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using AWS.Lambda.Powertools.Kafka.Avro;
 
 namespace AWS.Lambda.Powertools.Kafka.Tests
 {
