+++ version.json
@@ -1,18 +1,18 @@
 {
-	"Core": {
-		"Logging": "2.0.0",
-		"Metrics": "2.0.1",
-		"Tracing": "1.6.2",
-		"Metrics.AspNetCore": "0.1.0"
-	},
-	"Utilities": {
-		"Parameters": "1.3.1",
-		"Idempotency": "1.3.0",
-		"BatchProcessing": "1.2.1",
-		"EventHandler": "1.0.0",
-		"EventHandler.Resolvers.BedrockAgentFunction": "1.0.0",
-		"Kafka.Json" : "1.0.1",
-		"Kafka.Avro" : "1.0.1",
-		"Kafka.Protobuf" : "1.0.1"
-	}
+  "Core": {
+    "Logging": "2.0.0",
+    "Metrics": "2.0.1",
+    "Tracing": "1.6.2",
+    "Metrics.AspNetCore": "0.1.0"
+  },
+  "Utilities": {
+    "Parameters": "1.3.1",
+    "Idempotency": "1.3.0",
+    "BatchProcessing": "1.2.1",
+    "EventHandler": "1.0.0",
+    "EventHandler.Resolvers.BedrockAgentFunction": "1.0.0",
+    "Kafka.Json": "1.0.2",
+    "Kafka.Avro": "1.0.2",
+    "Kafka.Protobuf": "1.0.2"
+  }
 }
